finished laravel-admin

+++ database/migrations/2019_09_15_192440_create_articles_table.php
@@ -8,11 +8,11 @@ class CreatearticlesTable extends Migration
     public function up()
     {
         Schema::create('articles', function(Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('title')->index();
             $table->text('body');
             $table->bigInteger('user_id')->unsigned()->index();
-            $table->integer('category_id')->unsigned()->index();
+            $table->bigInteger('category_id')->unsigned()->index();
             $table->integer('reply_count')->unsigned()->default(0);
             $table->integer('view_count')->unsigned()->default(0);
             $table->integer('last_reply_user_id')->unsigned()->default(0);
